refactor(GN-157): 폴더 위치 이동
- 페이지 컴포넌트와 동일한 위치로 예약모달 컴포넌트를 이동
