+++ src/components/activity/ReservationCard/MobileReservationCard/ReservationFooterMenu.tsx
@@ -1,9 +1,9 @@
 import { useEffect } from 'react';
 
-import DateTimeSummary from '@/components/ActivityPage/ReservationCard/ReservationCommon/DateTimeSummary';
-import PriceDisplay from '@/components/ActivityPage/ReservationCard/ReservationCommon/PriceDisplay';
-import ReservationButton from '@/components/ActivityPage/ReservationCard/ReservationCommon/ReservationButton';
-import SelectDateTextButton from '@/components/ActivityPage/ReservationCard/ReservationCommon/SelectDateTextButton';
+import DateTimeSummary from '@/components/activity/ReservationCard/ReservationCommon/DateTimeSummary';
+import PriceDisplay from '@/components/activity/ReservationCard/ReservationCommon/PriceDisplay';
+import ReservationButton from '@/components/activity/ReservationCard/ReservationCommon/ReservationButton';
+import SelectDateTextButton from '@/components/activity/ReservationCard/ReservationCommon/SelectDateTextButton';
 import { Modal, useModal } from '@/components/common/Modal';
 import { useReservation } from '@/hooks/useReservation';
 import { isReservationValid } from '@/lib/utils/isReservationValid';
