+++ src/components/activity/ReservationCard/MobileReservationCard/Step1MobileCard.tsx
@@ -1,5 +1,5 @@
-import ReservationHeader from '@/components/ActivityPage/ReservationCard/ReservationCommon/ReservationHeader';
-import TimeSelector from '@/components/ActivityPage/ReservationCard/ReservationCommon/TimeSelector';
+import ReservationHeader from '@/components/activity/ReservationCard/ReservationCommon/ReservationHeader';
+import TimeSelector from '@/components/activity/ReservationCard/ReservationCommon/TimeSelector';
 import DatePicker from '@/components/common/DatePicker/DatePicker';
 import {
   CardEventHandlerType,
