+++ src/components/activity/ReservationCard/MobileReservationCard/index.tsx
