+++ src/pages/activity/[id].tsx
@@ -4,9 +4,9 @@ import CustomKebab from '@/components/activity/CustomKebab';
 import ImageGallery from '@/components/activity/ImageGallery';
 import KakaoMap from '@/components/activity/KakaoMap';
 import Location from '@/components/activity/Location';
+import ReservationCard from '@/components/activity/ReservationCard';
 import { ReviewRating } from '@/components/activity/Review';
 import ReviewList from '@/components/activity/ReviewList';
-import ReservationCard from '@/components/ActivityPage/ReservationCard';
 import Pagination from '@/components/common/Pagination';
 import useFetchData from '@/hooks/useFetchData';
 import { usePagination } from '@/hooks/usePagination';
